turned sticky compact view promo back on
