+++ _test-server/client/main.js
@@ -1,6 +1,6 @@
 import {bootstrap} from '../../main';
 const tracking = require('../../tracking');
 
-bootstrap({preset: 'complete'}, () => {
+bootstrap({preset: 'complete', compactViewPromo: true}, () => {
 	tracking.scrollDepth.init('n-ui-test');
 });
