+++ compact-view-promo/banner.html
@@ -1,0 +1,33 @@
+<div class="n-welcome-banner__wrapper">
+{{#with @root.welcomeBanner}}
+	<span class="n-welcome-banner__logo"></span>
+
+	<div class="n-welcome-banner__column n-welcome-banner__column--fill" data-welcome-banner-type="{{name}}">
+
+		<div class="n-welcome-banner__column n-welcome-banner__column--text">
+			<div class="n-welcome-banner__heading">{{title}}</div>
+			<p class="n-welcome-banner__strapline">{{strapline}}</p>
+		</div>
+
+		<p class="n-welcome-banner__column">
+			{{#with @root.welcomeBanner.ctas.primary}}
+				<a href="{{href}}" class="n-welcome-banner__button n-welcome-banner__button--primary" data-trackable="{{trackable}}" {{#if component-attr}}data-component="{{component-attr}}"{{/if}}>{{text}}</a>
+			{{/with}}
+		</p>
+
+		<div class="n-welcome-banner__column n-welcome-banner__column--actions">
+			{{#if anon.userIsLoggedIn}}
+				{{#if @root.flags.NextWelcomeHelp}}
+					<a href="http://help.ft.com/" data-trackable="help" class="n-welcome-banner__button--secondary">Help</a>
+				{{/if}}
+			{{/if}}
+			<a href="/feedback" data-trackable="nps-feedback" class="n-welcome-banner__link-action">Give feedback</a>
+		</div>
+	</div>
+	{{#if ../closable}}
+	<div class="n-welcome-banner__column">
+		<button data-trackable="toggled" data-action="welcome-banner-close" title="Close welcome banner" class="n-welcome-banner__button--close o-if-js if-fixed"></button>
+	</div>
+	{{/if}}
+{{/with}}
+</div>
