+++ compact-view-promo/pa11y-config.js
