+++ main.scss
@@ -18,5 +18,5 @@ $n-ui-foundations-applied: true;
 // TODO: move these components into n-ui
 $o-cookie-message-is-silent: false;
 @import 'o-cookie-message/main';
-@import 'welcome-message/main';
+@import 'compact-view-promo/main';
 @import 'subscription-offer-prompt/main';
